moving shit. decoupled mainnav, posts, postsgrid out of top
