+++ app/components/Post.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+export class Post extends React.Component {
+  render() {
+    return <h1>hey</h1>;
+  }
+}
